Use the source attribute to infer the repository URL
It is possible to use a fork of a Go project. In this case, the
repository URL is different from the import path. This is implement by
the "source" attribute in the Gopkg.toml constraint section.
So, if the source attribute exists, we use it instead of the import
path.